leptcam.c: general improvements
leptcam will now push imagery directly to a v4l2loopback device

Note: There are some stability improvements that are necessary
here.  The VoSPI interface (at least as tested on i.MX6) is not
reading fast enough to keep up with the interface on the Lepton.
When combined with image processing, it is sometimes causing
the Lepton to reset/resync the interface, which results in lower
framerate and decreased stability.